pyproject.toml
[project]
name = "filter-data-swamp"
version = "0.1.0"
description = "Filter data swamp pipeline using dlt and DuckDB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boring-semantic-layer[fastmcp,viz-altair,viz-plotly]>=0.1.12",
    "dbt-core>=1.10.13",
    "dbt-duckdb>=1.9.6",
    "dlt[dbt,duckdb,ducklake,filesystem,gs,parquet]>=1.12.1",
    "duckdb>=1.4.1",
    "ibis-framework[duckdb]>=11.0.0",
    "ipykernel>=7.0.1",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "rich>=13.0.0",
]
